# backend/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Smart Monitor API",
    "description": "Complete API documentation for Smart Monitor - Distributed System Monitoring Platform\n\n## Features\n- Agent Registration & Authentication\n- Real-time Metrics Streaming\n- System Monitoring (CPU, RAM, Disk)\n- Health & Readiness Checks\n\n## Quick Start\n1. Register agent: POST /v1/agent/register\n2. Stream metrics: POST /v1/stats/stream\n3. Query stats: GET /v1/stats/{hostname}\n\n## Authentication\nMost endpoints require access token obtained during registration.",
    "version": "1.0.0",
    "contact": {
      "name": "Smart Monitor Team",
      "email": "support@smartmonitor.example.com",
      "url": "https://smartmonitor.example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "access_token",
      "in": "query",
      "description": "Access token obtained from agent registration"
    },
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Bearer token format: 'Bearer {access_token}'"
    }
  },
...
  "paths": {
...
    "/search/stats": {
      "get": {
        "tags": ["Search"],
        "summary": "Search stats",
        "description": "Search stats documents (OpenSearch)",
        "operationId": "searchStats",
        "parameters": [
          {"name": "q", "in": "query", "type": "string", "description": "Free text query"},
          {"name": "hostname", "in": "query", "type": "string"},
          {"name": "limit", "in": "query", "type": "integer", "format": "int32", "default": 100}
        ],
        "responses": {
          "200": {"description": "Stats search result", "schema": {"$ref": "#/definitions/SearchStatsResponse"}},
          "500": {"description": "Server error", "schema": {"$ref": "#/definitions/ErrorResponse"}}
        },
        "security": [{"BearerAuth": []}]
      }
    },
    "/search/alerts": {
      "get": {
        "tags": ["Search"],
        "summary": "Search alerts",
        "description": "Search alerts with filters",
        "operationId": "searchAlerts",
        "parameters": [
          {"name": "hostname", "in": "query", "type": "string"},
          {"name": "severity", "in": "query", "type": "string", "enum": ["info", "warning", "critical"]},
          {"name": "status", "in": "query", "type": "string", "enum": ["open", "resolved"]},
          {"name": "limit", "in": "query", "type": "integer", "format": "int32", "default": 100}
        ],
        "responses": {
          "200": {"description": "Alerts list", "schema": {"$ref": "#/definitions/SearchAlertsResponse"}},
          "500": {"description": "Server error", "schema": {"$ref": "#/definitions/ErrorResponse"}}
        },
        "security": [{"BearerAuth": []}]
      }
    },
    "/search/events": {
      "get": {
        "tags": ["Search"],
        "summary": "Search events",
        "description": "Search events with filters",
        "operationId": "searchEvents",
        "parameters": [
          {"name": "q", "in": "query", "type": "string"},
          {"name": "hostname", "in": "query", "type": "string"},
          {"name": "type", "in": "query", "type": "string"},
          {"name": "level", "in": "query", "type": "string"},
          {"name": "limit", "in": "query", "type": "integer", "format": "int32", "default": 100},
          {"name": "offset", "in": "query", "type": "integer", "format": "int32", "default": 0}
        ],
        "responses": {
          "200": {"description": "Events list", "schema": {"$ref": "#/definitions/SearchEventsResponse"}},
          "500": {"description": "Server error", "schema": {"$ref": "#/definitions/ErrorResponse"}}
        },
        "security": [{"BearerAuth": []}]
      }
    },
...
    "/search/alerts/create": {
      "post": {
        "tags": ["Search"],
        "summary": "Create alert",
        "operationId": "createAlert",
        "parameters": [
          {"name": "body", "in": "body", "required": true, "schema": {"$ref": "#/definitions/AlertPayload"}}
        ],
        "responses": {
          "201": {"description": "Created", "schema": {"type": "object", "properties": {"id": {"type": "string"}}}},
          "400": {"description": "Bad request", "schema": {"$ref": "#/definitions/ErrorResponse"}},
          "500": {"description": "Server error", "schema": {"$ref": "#/definitions/ErrorResponse"}}
        },
        "security": [{"BearerAuth": []}]
      }
    },
    "/search/alerts/resolve": {
      "post": {
        "tags": ["Search"],
        "summary": "Resolve alert",
        "operationId": "resolveAlert",
        "parameters": [
          {"name": "id", "in": "query", "required": true, "type": "string", "description": "Alert ID"}
        ],
        "responses": {
          "200": {"description": "Resolved"},
          "400": {"description": "Bad request", "schema": {"$ref": "#/definitions/ErrorResponse"}},
          "500": {"description": "Server error", "schema": {"$ref": "#/definitions/ErrorResponse"}}
        },
        "security": [{"BearerAuth": []}]
      }
    },
    "/search/events/log": {
      "post": {
        "tags": ["Search"],
        "summary": "Log event",
        "operationId": "logEvent",
        "parameters": [
          {"name": "body", "in": "body", "required": true, "schema": {"$ref": "#/definitions/EventPayload"}}
        ],
        "responses": {
          "201": {"description": "Logged"},
          "400": {"description": "Bad request", "schema": {"$ref": "#/definitions/ErrorResponse"}},
          "500": {"description": "Server error", "schema": {"$ref": "#/definitions/ErrorResponse"}}
        },
        "security": [{"BearerAuth": []}]
      }
    },
...
  },
  "definitions": {
...
    "SearchStatsResponse": {
      "type": "object",
      "properties": {
        "total": {"type": "integer", "format": "int32"},
        "limit": {"type": "integer", "format": "int32"},
        "query": {"type": "string"},
        "result": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "hostname": {"type": "string"},
              "cpu": {"type": "number", "format": "double"},
              "ram": {"type": "number", "format": "double"},
              "disk": {"type": "number", "format": "double"},
              "timestamp": {"type": "integer", "format": "int64"}
            }
          }
        }
      }
    },
    "SearchAlertsResponse": {
      "type": "object",
      "properties": {
        "total": {"type": "integer", "format": "int32"},
        "limit": {"type": "integer", "format": "int32"},
        "hostname": {"type": "string"},
        "severity": {"type": "string"},
        "status": {"type": "string"},
        "result": {
          "type": "array",
          "items": {"$ref": "#/definitions/AlertPayload"}
        }
      }
    },
    "SearchEventsResponse": {
      "type": "object",
      "properties": {
        "total": {"type": "integer", "format": "int32"},
        "limit": {"type": "integer", "format": "int32"},
        "offset": {"type": "integer", "format": "int32"},
        "hostname": {"type": "string"},
        "type": {"type": "string"},
        "level": {"type": "string"},
        "result": {
          "type": "array",
          "items": {"$ref": "#/definitions/EventPayload"}
        }
      }
    },
    "AlertPayload": {
      "type": "object",
      "required": ["hostname", "alert_type", "severity", "title", "description"],
      "properties": {
        "id": {"type": "string"},
        "hostname": {"type": "string"},
        "alert_type": {"type": "string"},
        "severity": {"type": "string", "enum": ["critical", "high", "medium", "low"]},
        "title": {"type": "string"},
        "description": {"type": "string"},
        "timestamp": {"type": "integer", "format": "int64"},
        "status": {"type": "string", "enum": ["active", "resolved"]},
        "value": {"type": "number", "format": "double"},
        "threshold": {"type": "number", "format": "double"},
        "metadata": {"type": "object", "additionalProperties": {"type": "string"}}
      }
    },
    "EventPayload": {
      "type": "object",
      "required": ["hostname", "event_type", "event_name", "message", "source", "level"],
      "properties": {
        "id": {"type": "string"},
        "hostname": {"type": "string"},
        "event_type": {"type": "string"},
        "event_name": {"type": "string"},
        "timestamp": {"type": "integer", "format": "int64"},
        "message": {"type": "string"},
        "source": {"type": "string"},
        "user": {"type": "string"},
        "process_id": {"type": "integer", "format": "int32"},
        "process_name": {"type": "string"},
        "level": {"type": "string", "enum": ["info", "warning", "error", "critical"]},
        "details": {"type": "object", "additionalProperties": {}}
      }
    },
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": false
        },
        "message": {
          "type": "string",
          "description": "Error description",
          "example": "Invalid request parameters"
        },
        "code": {
          "type": "string",
          "description": "Error code",
          "enum": [
            "INVALID_REQUEST",
            "UNAUTHORIZED",
            "NOT_FOUND",
            "VALIDATION_ERROR",
            "INTERNAL_ERROR"
          ],
          "example": "INVALID_REQUEST"
        },
        "details": {
          "type": "object",
          "description": "Additional error details"
        }
      }
    }
  },
...
}